requirements.txt
Flask
flask_jwt_extended==4.6.0
flask_restful==0.3.10
pydantic_settings==2.2.1
PyJWT==2.8.0
pymongo==3.13.0
python-dotenv==1.0.1
requests
SQLAlchemy==2.0.4
sqlalchemy_json==0.7.0
torch==2.3.0
transformers==4.40.2